debateai/requirements.txt
crewai
pydantic
python-dotenv
crewai[google-genai]
